Aug 29, 2023, 7:25 AM

+++ ch_06/q14.java
@@ -10,7 +10,7 @@ public class q14 {
     }
     
     public static double estimatePI(int i) {
-        double m = Math.pow(-1, i + 1) / (2 * i - 1);
+        double m = (double) Math.pow(-1, i + 1) / (2 * i - 1);
         return 4 * m;
     }
 }
